test(scripts): extract growth-dashboard arg parser and cover it
scripts/growth-dashboard.mjs parsed its --json/--gsc/--limit flags with a local
parseArgs that had no tests. Hoist it into the existing scripts/lib/growth-dashboard.mjs
(which the script already imports for buildWorklist) as parseGrowthDashboardArgs
and use it from the script.

- scripts/lib/growth-dashboard.mjs: add parseGrowthDashboardArgs.
- scripts/growth-dashboard.mjs: import the shared parser; drop the local copy.
- tests/growth-dashboard.test.ts: add cases for the flag defaults, --json plus
  --gsc and --limit, the '' / 50 fallbacks for valueless --gsc and non-numeric or
  zero --limit, and ignored tokens.

No behavior change; the CLI arguments parse identically.
